Extract peer-link wire enums into wire.py
Second PR in the peer_link split arc (plan:
peer_link_split.md). Moves TerminateReason and AppMessageType
out of peer_link/__init__.py into peer_link/wire.py
byte-for-byte. __init__.py re-exports both via
"from .wire import AppMessageType, TerminateReason" so
external callers (receiver.py, submit_job.py,
artifacts_download.py, job_fanout.py, peer_link_client/) keep
using "from .peer_link import TerminateReason" / etc.
without change.

Risk: LOW. Pure StrEnums; no state, no test patches, no
runtime behaviour. The __all__ pin guarantees the re-export
survives a future ruff sweep that might prune the import
otherwise.

__init__.py: 1079 -> 977 lines (-102). wire.py: 110 lines.

All 3242 non-e2e tests pass; ruff + mypy clean.
